refactor : TimeManager로 이름 정정
