+++ src/main/java/roomescape/global/config/TimeConfig.java
@@ -2,14 +2,14 @@ package roomescape.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import roomescape.global.time.SystemTimeProvider;
-import roomescape.global.time.TimeProvider;
+import roomescape.global.time.SystemTimeManager;
+import roomescape.global.time.TimeManager;
 
 @Configuration
 public class TimeConfig {
 
     @Bean
-    public TimeProvider timeProvider() {
-        return new SystemTimeProvider();
+    public TimeManager timeProvider() {
+        return new SystemTimeManager();
     }
 }
