+++ src/main/java/roomescape/global/time/TimeManager.java
@@ -1,0 +1,9 @@
+package roomescape.global.time;
+
+import java.time.LocalDateTime;
+
+public interface TimeManager {
+
+    LocalDateTime nowDateTime();
+
+}
