--- src/main/java/roomescape/global/time/TimeProvider.java
@@ -1,14 +1,0 @@
-package roomescape.global.time;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-public interface TimeProvider {
-
-    LocalDateTime nowDateTime();
-
-    LocalDate nowDate();
-
-    LocalTime nowTime();
-}
