+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.global.time.TimeProvider;
+import roomescape.global.time.TimeManager;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.exception.ReservationAlreadyExistsException;
 import roomescape.reservation.exception.ReservationResourceNotFoundException;
@@ -22,7 +22,7 @@ public class ReservationService {
 
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
-    private final TimeProvider timeProvider;
+    private final TimeManager timeManager;
 
     public List<ReservationResult> getAll() {
         return reservationRepository.findAll();
@@ -39,7 +39,7 @@ public class ReservationService {
         validateDuplicate(date, timeId);
 
         Reservation reservation = Reservation.createNew(name, date, reservationTime.getId());
-        reservation.validateNotPast(reservationTime.getStartAt(), timeProvider.nowDateTime());
+        reservation.validateNotPast(reservationTime.getStartAt(), timeManager.nowDateTime());
 
         Reservation savedReservation = reservationRepository.save(reservation);
 
@@ -67,7 +67,7 @@ public class ReservationService {
         validateDuplicate(date, timeId);
 
         reservation = reservation.modify(date, reservationTime.getId());
-        reservation.validateNotPast(reservationTime.getStartAt(), timeProvider.nowDateTime());
+        reservation.validateNotPast(reservationTime.getStartAt(), timeManager.nowDateTime());
 
         reservationRepository.update(reservation);
     }
