+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.global.time.TimeProvider;
+import roomescape.global.time.TimeManager;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.exception.ReservationAlreadyExistsException;
 import roomescape.reservation.exception.ReservationPastDateException;
@@ -34,7 +34,7 @@ class ReservationServiceTest {
     private JdbcTemplate jdbcTemplate;
 
     @Autowired
-    private TimeProvider timeProvider;
+    private TimeManager timeManager;
 
     private ReservationService reservationService;
     private ReservationRepository reservationRepository;
@@ -47,7 +47,7 @@ class ReservationServiceTest {
         reservationTimeRepository = new JdbcReservationTimeRepository(jdbcTemplate);
         themeRepository = new JdbcThemeRepository(jdbcTemplate);
 
-        reservationService = new ReservationService(reservationRepository, reservationTimeRepository, timeProvider);
+        reservationService = new ReservationService(reservationRepository, reservationTimeRepository, timeManager);
     }
 
     @Test
