+++ src/test/java/roomescape/support/time/FixedTimeManager.java
@@ -1,0 +1,19 @@
+package roomescape.support.time;
+
+import java.time.LocalDateTime;
+import roomescape.global.time.TimeManager;
+
+public class FixedTimeManager implements TimeManager {
+
+    private LocalDateTime current;
+
+    public FixedTimeManager(final LocalDateTime current) {
+        this.current = current;
+    }
+
+    @Override
+    public LocalDateTime nowDateTime() {
+        return current;
+    }
+
+}
